refactor: convert Map to function component
