+++ src/components/Map.js
@@ -1,0 +1,26 @@
+import React, { useEffect, useRef } from "react"
+
+export default function Map({ options, onMount, className }) {
+  const props = { ref: useRef(), className }
+  const onLoad = () => {
+    const map = new window.google.maps.Map(props.ref.current, options)
+    onMount && onMount(map)
+  }
+
+  useEffect(() => {
+    if (!window.google) {
+      const script = document.createElement(`script`)
+      script.type = `text/javascript`
+      script.src = `https://maps.google.com/maps/api/js?key=${
+        process.env.GATSBY_GOOGLE_MAPS_API_KEY
+      }`
+      const headScript = document.getElementsByTagName(`script`)[0]
+      headScript.parentNode.insertBefore(script, headScript)
+      script.addEventListener(`load`, onLoad)
+      return () => script.removeEventListener(`load`, onLoad)
+    } else onLoad()
+  })
+
+  const mapCss = `height: 70vh; margin: 1em 0; border-radius: 0.5em;`
+  return <div css={mapCss} {...props} />
+}
