--- src/components/Map/styles.js
@@ -1,6 +1,0 @@
-import styled from 'styled-components'
-
-export const MapContainer = styled.div`
-  height: 70vh;
-  ${props => props.css};
-`
